+++ firmware/micromound-c/include/mm_version.h
@@ -6,6 +6,6 @@
 #ifndef MM_VERSION_H
 #define MM_VERSION_H
 
-#define MM_VERSION "0.9.31"
+#define MM_VERSION "0.9.32"
 
 #endif